+++ package.json
@@ -1,71 +1,25 @@
 {
-  "name": "groqd",
-  "version": "0.12.0",
-  "license": "MIT",
-  "author": {
-    "name": "Formidable",
-    "url": "https://formidable.com"
+  "name": "groqd-workspace",
+  "private": "true",
+  "workspaces": {
+    "packages": [
+      "packages/*"
+    ]
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/formidablelabs/groqd.git"
-  },
-  "homepage": "https://github.com/formidablelabs/groqd",
-  "keywords": [
-    "sanity",
-    "groq",
-    "query",
-    "typescript"
-  ],
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "exports": {
-    ".": [
-      {
-        "import": "./dist/index.mjs",
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      },
-      "./dist/index.js"
-    ],
-    "./package.json": "./package.json"
-  },
-  "files": [
-    "dist"
-  ],
   "scripts": {
-    "test:watch": "vitest",
-    "test": "vitest run",
-    "typecheck": "tsc --noEmit",
-    "lint": "eslint src --quiet",
-    "check:ci": "yarn typecheck && yarn lint && yarn test",
-    "build": "tsup",
-    "preversion": "yarn check:ci",
-    "prepare": "yarn build",
-    "changeset": "changeset",
-    "version": "yarn changeset version && yarn install --frozen-lockfile"
+    "lint": "eslint packages/**/src --quiet",
+    "test": "yarn workspaces run test",
+    "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
-    "@sanity/client": "^3.4.1",
     "@svitejs/changesets-changelog-github-compact": "^1.1.0",
     "@typescript-eslint/eslint-plugin": "^5.52.0",
     "@typescript-eslint/parser": "^5.52.0",
     "eslint": "^8.34.0",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "groq-js": "^1.1.1",
     "prettier": "^2.8.4",
-    "tiny-invariant": "^1.3.1",
-    "tsup": "^6.3.0",
     "typescript": "^4.9.5",
-    "vitest": "^0.28.5",
     "@changesets/cli": "^2.26.0"
-  },
-  "dependencies": {
-    "zod": "^3.20.2"
-  },
-  "engines": {
-    "node": ">= 14"
   }
 }
