+++ packages/groqd/src/contentBlock.ts
