+++ packages/groqd/src/makeSafeQueryRunner.ts
