+++ packages/groqd/src/sanityImage.test.ts
