+++ packages/groqd/src/selectionUtils.ts
